Add initial Protobuf files
The Pulumi runtime requires gRPC/Protobufs in order to communicate
between the language host and the engine. This commits the generated
files, using github.com/awakesecurity/gRPC-haskell's compile-proto-file
utility. There is, sadly, some hacky sed'ding required because of
the module layout that tool produces (with, apparently, no way to
override).