Launch programs using 'cabal exec'
Right now we're letting you run the Pulumi user program through
'runhaskell' to avoid needing an extra compilation step. (We likely
want to enable that down the road, perhaps as an opt-in.) Unfortunately,
the way we execute 'runhaskell' means you can't really have dependencies
very easily and indeed there was a hack to link with the Pulumi SDK.

This changes to use 'cabal exec runhaskell -- ...' instead. This
respects the Cabal dependencies and lets you more conveniently depend
on other libraries. It also cleans up the Pulumi SDK hack since we
can use a Cabal project file.